fix compile error

+++ src/server/raftkv/raft_extension.rs
@@ -114,7 +114,7 @@ where
         region_epoch: RegionEpoch,
         split_keys: Vec<Vec<u8>>,
         source: String,
-        _encrypt: u16,
+        _encrypt: u32,
     ) -> BoxFuture<'static, kv::Result<Vec<Region>>> {
         let (cb, rx) = paired_future_callback();
         let req = CasualMessage::SplitRegion {
